pyproject.toml
[project]
name = "small-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.123.10",
    "google-adk>=1.21.0",
    "litellm>=1.80.9",
    "load-dotenv>=0.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
